简化同步流程

+++ src/main/java/com/xinshijie/gallery/dto/UserCollectionDto.java
@@ -40,7 +40,7 @@ public class UserCollectionDto implements Serializable {
     @Size(max = 100, message = " 超出最大长度 100")
     private String title;
 
-    private Long aid;
+    private Integer aid;
 
     private Integer ctype;
